Removed source from diff view
