+++ mlflow/server/js/src/experiment-tracking/components/ExperimentView.js
@@ -49,7 +49,6 @@ import {
   DEFAULT_START_TIME,
   ATTRIBUTE_COLUMN_SORT_LABEL,
   ATTRIBUTE_COLUMN_SORT_KEY,
-  ATTRIBUTE_COLUMN_LABELS,
   COLUMN_SORT_BY_ASC,
   COLUMN_SORT_BY_DESC,
   SORT_DELIMITER_SYMBOL,
@@ -1151,7 +1150,7 @@ export class ExperimentView extends Component {
   handleDiffSwitchChange() {
     this.setState({ diffSwitchSelected: !this.state.diffSwitchSelected }, () => {
       const categorizedUncheckedKeys = this.state.diffSwitchSelected
-        ? this.getCategorizedColumnsDiffView()
+        ? ExperimentViewUtil.getCategorizedColumnsDiffView(this.props)
         : {
             [COLUMN_TYPES.ATTRIBUTES]: [],
             [COLUMN_TYPES.PARAMS]: [],
@@ -1160,74 +1159,6 @@ export class ExperimentView extends Component {
           };
       this.handleColumnSelectionCheck(categorizedUncheckedKeys);
     });
-  }
-
-  /**
-   * Obtain the categorized columns for which the values in them
-   * have only a single value (or are undefined)
-   */
-  getCategorizedColumnsDiffView() {
-    const { paramKeyList, metricKeyList, runInfos, paramsList, metricsList, tagsList } = this.props;
-    const tagKeyList = Utils.getVisibleTagKeyList(tagsList);
-    const attributeKeyList = [
-      ATTRIBUTE_COLUMN_LABELS.RUN_NAME,
-      ATTRIBUTE_COLUMN_LABELS.USER,
-      ATTRIBUTE_COLUMN_LABELS.SOURCE,
-      ATTRIBUTE_COLUMN_LABELS.VERSION,
-    ];
-    let attributes = [];
-    let params = [];
-    let metrics = [];
-    let tags = [];
-
-    for (let index = 0, n = runInfos.length; index < n; ++index) {
-      const paramsMap = ExperimentViewUtil.toParamsMap(paramsList[index]);
-      const metricsMap = ExperimentViewUtil.toMetricsMap(metricsList[index]);
-      const tagsMap = tagsList[index];
-
-      attributes.push([
-        Utils.getRunName(tagsList[index]),
-        Utils.getUser(runInfos[index], tagsList[index]),
-        Utils.formatSource(tagsList[index]),
-        Utils.getSourceVersion(tagsList[index]),
-      ]);
-      params.push(
-        paramKeyList.map((paramKey) => {
-          return paramsMap[paramKey] ? paramsMap[paramKey].getValue() : '';
-        }),
-      );
-      metrics.push(
-        metricKeyList.map((metricKey) => {
-          return metricsMap[metricKey] ? metricsMap[metricKey].getValue() : '';
-        }),
-      );
-      tags.push(
-        tagKeyList.map((tagKey) => {
-          return tagsMap[tagKey] ? tagsMap[tagKey].getValue() : '';
-        }),
-      );
-    }
-    // Transpose the matrices so that we can evaluate the values 'column-based'
-    attributes = _.unzip(attributes);
-    params = _.unzip(params);
-    metrics = _.unzip(metrics);
-    tags = _.unzip(tags);
-    const allEqual = (arr) => arr.every((val) => val === arr[0]);
-
-    return {
-      [COLUMN_TYPES.ATTRIBUTES]: attributeKeyList.filter((v, index) => {
-        return allEqual(attributes[index]);
-      }),
-      [COLUMN_TYPES.PARAMS]: paramKeyList.filter((v, index) => {
-        return allEqual(params[index]);
-      }),
-      [COLUMN_TYPES.METRICS]: metricKeyList.filter((v, index) => {
-        return allEqual(metrics[index]);
-      }),
-      [COLUMN_TYPES.TAGS]: tagKeyList.filter((v, index) => {
-        return allEqual(tags[index]);
-      }),
-    };
   }
 
   onSearch(e, searchInput) {
